+++ wp-cli-sync.php
@@ -2,7 +2,7 @@
 /*
 Plugin Name:  WP-CLI Sync
 Description:  A WP-CLI command for syncing a live site to a development environment
-Version:      1.3.0
+Version:      1.3.1
 Author:       Jon Beaumont-Pike
 Author URI:   https://jonbp.co.uk/
 License:      MIT License
@@ -60,6 +60,9 @@ if ( defined( 'WP_CLI' ) && WP_CLI ) {
     $ssh_username = $_ENV['LIVE_SSH_USERNAME'];
     $rem_proj_loc = $_ENV['REMOTE_PROJECT_LOCATION'];
 
+    // Welcome
+    task_message('Running .env file and connection checks...', 'WP-CLI Sync', 97);
+
     /**
      * BEGIN VAR / CONNECTION CHECKS
      */
@@ -68,7 +71,7 @@ if ( defined( 'WP_CLI' ) && WP_CLI ) {
     if (empty($ssh_hostname) || empty($ssh_username) || empty($rem_proj_loc)) {
 
       // Exit Messages
-      task_message('some/all dev sync vars are not set in .env file', 'Error', 31);
+      task_message('some/all dev sync vars are not set in .env file', 'Error', 31, false);
 
       // Line Break + Color Reset + Exit
       lb_cr();
@@ -80,7 +83,8 @@ if ( defined( 'WP_CLI' ) && WP_CLI ) {
     if(($rem_proj_loc[0] != '/') && ($rem_proj_loc[0] != '~')) {
 
       // Exit Messages
-      task_message('Incorrect formatting of the REMOTE_PROJECT_LOCATION variable', 'Error', 31);
+      task_message('Incorrect formatting of the REMOTE_PROJECT_LOCATION variable', 'Error', 31, false);
+      task_message('Ensure that the path begins with either / or ~/', 'Hint', 33);
 
       // Line Break + Color Reset + Exit
       lb_cr();
@@ -91,7 +95,8 @@ if ( defined( 'WP_CLI' ) && WP_CLI ) {
       if($rem_proj_loc[1] != '/') {
 
         // Exit Messages
-        task_message('Incorrect formatting of the REMOTE_PROJECT_LOCATION variable', 'Error', 31);
+        task_message('Incorrect formatting of the REMOTE_PROJECT_LOCATION variable', 'Error', 31, false);
+        task_message('Ensure that the path begins with either / or ~/', 'Hint', 33);
 
         // Line Break + Color Reset + Exit
         lb_cr();
@@ -108,7 +113,8 @@ if ( defined( 'WP_CLI' ) && WP_CLI ) {
     if ($live_server_status == '255') {
 
       // Exit Messages
-      task_message('Cannot connect to live server over SSH', 'Error', 31);
+      task_message('Cannot connect to live server over SSH', 'Error', 31, false);
+      task_message('Check that your LIVE_SSH_HOSTNAME and LIVE_SSH_USERNAME variables are correct', 'Hint', 33);
 
       // Line Break + Color Reset + Exit
       lb_cr();
@@ -123,14 +129,17 @@ if ( defined( 'WP_CLI' ) && WP_CLI ) {
     if ($live_server_check == 'false') {
 
       // Exit Messages
-      task_message('Connected but cannot find remote WP-CLI', 'Error', 31);
-      task_message('Either WP-CLI Sync is not installed on the live server or the path to your project is incorrect', 'Notice');
+      task_message('Connected but cannot find remote WP-CLI', 'Error', 31, false);
+      task_message('Either WP-CLI Sync is not installed on the live server or the REMOTE_PROJECT_LOCATION variable is incorrect', 'Hint', 33);
 
       // Line Break + Color Reset + Exit
       lb_cr();
       exit();
 
     }
+
+    // Checks Success
+    task_message('Running sync...', 'Connected', 32, false);
 
     // Plugin Vars
     $dev_activated_plugins = $_ENV['DEV_ACTIVATED_PLUGINS'];